CC-684 Edited properties lesson tests files based on the proof library changes
