+++ answers/Properties/84334020-7682-4b7d-bfda-e7d45384630b/ManipulateObjectPropertiesTest.php
