+++ answers/Properties/e9640e3b-0b32-4232-95cd-98a3647036a6/MissingDollarSignOnObjectTest.php
