+++ tests/Properties/84334020-7682-4b7d-bfda-e7d45384630b/ManipulateObjectPropertiesTest.php
@@ -1,8 +1,15 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class ManipulateObjectPropertiesTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class ManipulateObjectPropertiesTest extends TestCase
 {
+    protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+    }
 
     public function testPhpStartTag()
     {
