+++ tests/Properties/88ff20f9-2d23-4325-b1d1-520a29351a30/CorrectMultipleErrorsTest.php
@@ -1,8 +1,15 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class CorrectMultipleErrorsTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class CorrectMultipleErrorsTest extends TestCase
 {
+    protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+    }
 
     public function testPhpStartTag()
     {
