+++ tests/Properties/e9640e3b-0b32-4232-95cd-98a3647036a6/MissingDollarSignOnObjectTest.php
@@ -1,8 +1,15 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class RemoveUnpexpectedDollarSignTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class MissingDollarSignOnObjectTest extends TestCase
 {
+    protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+    }
 
     public function testPhpStartTag()
     {
@@ -73,7 +80,7 @@ class RemoveUnpexpectedDollarSignTest extends Proof
         $address = $subNodes->find('property[name="address", type="public"]');
 
         $this->assertEquals(1, $address->count(), "Expecting a public class property named 'address'.");
-    }  
+    }   
 
     public function testNameValue()
     {
